requirements.txt
lxml
websocket-client
numpy
gensim
scipy
nltk
sklearn
pandas
stop_words
kaggle
